#72 - refactor : Twitch GameStreamData 관련 코드 레포지토리 분리
