+++ core/data/src/main/java/com/hegunhee/data/di/RepositoryModule.kt
@@ -1,8 +1,10 @@
 package com.hegunhee.data.di
 
 import com.hegunhee.data.repository.DefaultRepository
+import com.hegunhee.data.repository.twitch.DefaultTwitchGameStreamRepository
 import com.hegunhee.data.repository.twitch.DefaultTwitchStreamRepository
 import com.hegunhee.domain.repository.Repository
+import com.hegunhee.domain.repository.twitch.TwitchGameStreamRepository
 import com.hegunhee.domain.repository.twitch.TwitchStreamRepository
 import dagger.Binds
 import dagger.Module
@@ -21,4 +23,8 @@ abstract class RepositoryModule {
     @Binds
     @Singleton
     abstract fun provideTwitchStreamRepository(defaultTwitchStreamRepository: DefaultTwitchStreamRepository): TwitchStreamRepository
+
+    @Binds
+    @Singleton
+    abstract fun provideTwitchGameStreamRepository(defaultTwitchGameStreamRepository: DefaultTwitchGameStreamRepository): TwitchGameStreamRepository
 }
