+++ core/domain/src/main/java/com/hegunhee/domain/repository/twitch/TwitchGameStreamRepository.kt
@@ -1,0 +1,8 @@
+package com.hegunhee.domain.repository.twitch
+
+import com.hegunhee.domain.model.twitch.StreamDataType
+
+interface TwitchGameStreamRepository {
+
+    suspend fun getGameStreamDataList(gameId: String): Result<List<StreamDataType.OnlineData>>
+}
