+++ core/domain/src/main/java/com/hegunhee/domain/usecase/twitch/GetGameStreamDataListUseCase.kt
@@ -1,12 +1,12 @@
 package com.hegunhee.domain.usecase.twitch
 
 import com.hegunhee.domain.model.twitch.StreamDataType
-import com.hegunhee.domain.repository.Repository
+import com.hegunhee.domain.repository.twitch.TwitchGameStreamRepository
 import javax.inject.Inject
 
-class GetGameStreamDataListUseCase @Inject constructor(private val repository : Repository){
+class GetGameStreamDataListUseCase @Inject constructor(private val gameStreamRepository: TwitchGameStreamRepository){
 
     suspend operator fun invoke(gameId : String) : Result<List<StreamDataType.OnlineData>>{
-        return repository.getGameStreamDataList(gameId)
+        return gameStreamRepository.getGameStreamDataList(gameId)
     }
 }
